refact(api): move worker handlers
