+++ internal/daemon/worker/worker_proxy_service.go
@@ -1,4 +1,4 @@
-package workers
+package worker
 
 import (
 	"context"
